+++ src/Components/Configuration/EditCard/Entity.tsx
